\name{as.data.frame-methods}
\docType{methods}
\alias{as.data.frame-methods}
\title{ \code{as.data.frame} Returning content of data.frame.}
\description{Methods for function \code{as.data.frame}}
\section{Methods}{
\describe{
\item{\code{signature(x = "gapSites")}}{Method for 'gapSites'.} 
\item{\code{signature(x = "annGapSites")}}{Method for 'annGapSites'.}
}}
\keyword{methods}
\keyword{as.data.frame}
