\name{c-methods}
\docType{methods}
\alias{c-methods}
\title{ Coercing functions \code{c}. }
\description{Coerce objects by binding contained data.}
\section{Methods}{
\describe{
\item{\code{signature(x = "cRanges")}}{Method for 'cRanges'.}
\item{\code{signature(x = "gapSites")}}{Method for 'gapSites'.}
}}
\keyword{methods}
\keyword{c}
