\name{dim-methods}
\docType{methods}
\alias{dim-methods}
\title{ \code{dim}: Return dimensions of contained data.frame.}
\description{Methods for function \code{dim}}
\section{Methods}{
\describe{
\item{\code{signature(x = "cRanges")}}{Method for 'cRanges'.}
\item{\code{signature(x = "gapSites")}}{Method for 'gapSites'.}
}}
\keyword{methods}
\keyword{dim}
