\name{head-methods}
\docType{methods}
\alias{head-methods}
\title{\code{head} Return first lines of contained data.frame.}
\description{Methods for function \code{head}.}
\section{Methods}{
\describe{
\item{\code{signature(x = "cRanges")}}{Method for 'cRanges'.}
\item{\code{signature(x = "aaGapSites")}}{Method for 'aaGapSites'.}
\item{\code{signature(x = "cdRanges")}}{Method for 'cdRanges'.}
\item{\code{signature(x = "cRanges")}}{Method for 'cRanges'.}
\item{\code{signature(x = "dnaGapSites")}}{Method for 'dnaGapSites'.}
\item{\code{signature(x = "gapSites")}}{Method for 'gapSites'.}
}}
\keyword{methods}
\keyword{head}
