\name{initialize-methods}
\docType{methods}
\alias{initialize-methods}
\title{ \code{initialize} Initializing objects.}
\description{Methods for function \code{initialize}}
\section{Methods}{
\describe{
\item{\code{signature(.Object = "cdRanges")   }}{Method for 'cdRanges'.}
\item{\code{signature(.Object = "cRanges")    }}{Method for 'cRanges'.}
\item{\code{signature(.Object = "keyProfiler")}}{Method for 'keyProfiler'.}
\item{\code{signature(.Object = "SpliceCountSet")}}{Method for 'SpliceCountSet'.}
}}
\keyword{methods}
\keyword{initialize}
