\name{merge-methods}
\docType{methods}
\alias{merge-methods}
\title{\code{merge} Merging two objects into one.}
\description{Methods for function \code{merge}}
\section{Methods}{
\describe{
\item{\code{signature(x = "gapSites", y = "ANY")}}{Method for 'gapSites'.}
}}
\keyword{methods}
\keyword{merge}
