\name{sortTable-methods}
\docType{methods}
\alias{sortTable-methods}
\title{ Sorting contained data with \code{sortTable}. }
\description{Sorting tables by key columns.}
\section{Methods}{
\describe{
\item{\code{signature(x = "cRanges")}}{Method for 'cRanges'. Key columns: seqid, start, end}
\item{\code{signature(x = "gapSites")}}{Method for 'gapSites'. Key columns: seqid, lend, rstart}
}}
\keyword{methods}
\keyword{sortTables}
