\name{SpliceCountSet-class}
\Rdversion{1.1}
\docType{class}
\alias{SpliceCountSet-class}
\alias{initialize,SpliceCountSet-method}
\title{Class \code{"SpliceCountSet"}}
\description{
Directly inherits from ExpressionSet
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("SpliceCountSet", ...)}.
}
\author{
Wolfgang Kaisers
}
\examples{
# scs<-new("SpliceCountSet")
}
\keyword{SpliceCountSet}
