\name{[-methods}
\docType{methods}
\alias{[-methods}
\title{ Methods for Function \code{[}.}
\description{Methods for function \code{[}}
\section{Methods}{
\describe{
\item{\code{signature(x = "cRanges",i="ANY",j="ANY",drop="ANY")}}{Method for 'cRanges'.}
\item{\code{signature(x = "gapSites",i="ANY",j="ANY",drop="ANY")}}{Method for 'gapSites.'}
}}
\keyword{methods}
